+++ backend/src/main/java/br/puccamp/listbuy/dao/ProductDAO.java
@@ -11,20 +11,20 @@ import java.util.ArrayList;
 
 public class ProductDAO extends GenericDAO {
 
-    public Produtos listarInformacoesProdutos (int idProduto) {
+    public Produtos listarInformacoesProdutos (Long idProduto) {
         StringBuilder sql = new StringBuilder();
         sql.append("SELECT * FROM PRODUTOS ");
         sql.append("WHERE ID_PRODUTO = ?");
         try (Connection connection = getConnection()) {
             PreparedStatement stmt = connection.prepareStatement(sql.toString());
-            stmt.setInt(1, idProduto);
+            stmt.setLong(1, idProduto);
             ResultSet rs = stmt.executeQuery();
             Produtos retorno = new Produtos();
             if (rs.next()) {
-                retorno.setId_produto(rs.getInt(1));
+                retorno.setId_produto(rs.getLong(1));
                 retorno.setNome(rs.getString(2));
                 retorno.setDescricao(rs.getString(3));
-                retorno.setId_categoria(rs.getInt(4));
+                retorno.setId_categoria(rs.getLong(4));
                 retorno.setAtivo(rs.getString(5));
                 return retorno;
             }
